+++ fj-doc-val-p7m/src/main/java/org/fugerit/java/doc/val/p7m/NopOutputStream.java
@@ -1,0 +1,22 @@
+package org.fugerit.java.doc.val.p7m;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+public class NopOutputStream extends OutputStream {
+
+    @Override
+    public void write(byte[] b) throws IOException {
+        // do nothing implementation
+    }
+
+    @Override
+    public void write(byte[] b, int off, int len) throws IOException {
+        // do nothing implementation
+    }
+
+    @Override
+    public void write(int b) throws IOException {
+        // do nothing implementation
+    }
+}